+++ hw/mips/mips_ls3a.c
@@ -45,6 +45,7 @@
 #include "hw/pci/pci_host.h"
 #include "hw/ssi/ssi.h"
 #include "loongson_bootparam.h"
+#include "target/mips/internal.h"
 
 #ifdef DEBUG_LS3A
 #define DPRINTF(fmt, ...) \
@@ -145,8 +146,8 @@ static int set_bootparam(ram_addr_t initrd_offset,long initrd_size)
 	*parg_env++=0;
 
 	//env
-	sprintf(memenv,"memsize=%d",loaderparams.ram_size>0x0f000000?240:(loaderparams.ram_size>>20));
-	sprintf(highmemenv,"highmemsize=%d",loaderparams.ram_size>0x10000000?(loaderparams.ram_size>>20)-256:0);
+	sprintf(memenv,"memsize=%d",(int)(loaderparams.ram_size>0x0f000000?240:(loaderparams.ram_size>>20)));
+	sprintf(highmemenv,"highmemsize=%d",(int)(loaderparams.ram_size>0x10000000?(loaderparams.ram_size>>20)-256:0));
 
 
 	for(i=0;i<sizeof(pmonenv)/sizeof(char *);i++)
@@ -213,8 +214,8 @@ static int set_bootparam1(ram_addr_t initrd_offset,long initrd_size)
 
 	//env
 
-	sprintf(memenv,"%d",loaderparams.ram_size>0x0f000000?240:(loaderparams.ram_size>>20));
-	sprintf(highmemenv,"%d",loaderparams.ram_size>0x10000000?(loaderparams.ram_size>>20)-256:0);
+	sprintf(memenv,"%d",(int)(loaderparams.ram_size>0x0f000000?240:(loaderparams.ram_size>>20)));
+	sprintf(highmemenv,"%d",(int)(loaderparams.ram_size>0x10000000?(loaderparams.ram_size>>20)-256:0));
 	setenv("memsize", memenv, 1);
 	setenv("highmemsize", highmemenv, 1);
